% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellmer-core.R
\name{capture}
\alias{capture}
\title{Capture chat model response with proper handling}
\usage{
capture(original_chat, prompt, type, eval_rounds, echo, ...)
}
\arguments{
\item{original_chat}{Original chat model object}

\item{prompt}{Prompt text}

\item{type}{Type specification for structured data}

\item{eval_rounds}{Number of evaluation rounds for structured data extraction resulting in refined data}
}
\value{
List containing response information
}
\description{
Capture chat model response with proper handling
}
\keyword{internal}
