% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\docType{package}
\name{chatlot-package}
\alias{chatlot}
\alias{chatlot-package}
\title{chatlot: Synchronous Batch Processing for Chat Models}
\description{
Synchronous process lots of large language model chats with 'ellmer'. Provides both sequential and parallel processing of chat interactions with features including tool calling and structured data extraction. Enables workflow management through progress tracking and recovery and automatic retry with backoff. Additional quality-of-life features include verbosity (or echo) control and sound notifications. Parallel processing is implemented via the 'future' framework. Includes methods for retrieving progress status, chat texts, and chat objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dylanpieper.github.io/chatlot/}
}

}
\author{
\strong{Maintainer}: Dylan Pieper \email{dylanpieper@gmail.com}

}
\keyword{internal}
