% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatlot-core.R
\name{create_results}
\alias{create_results}
\title{Create results object from lot}
\usage{
create_results(result)
}
\arguments{
\item{result}{Lot object}
}
\value{
Results object with class "lot"
}
\description{
Create results object from lot
}
\keyword{internal}
