% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellmer.R
\name{handle_batch_interrupt}
\alias{handle_batch_interrupt}
\title{Update batch state with new responses}
\usage{
handle_batch_interrupt(result, beep)
}
\arguments{
\item{result}{A batch object containing processing state}

\item{beep}{Logical indicating whether to play a sound}

\item{x}{A batch object
Handle batch interruption}
}
\value{
NULL (called for side effects)
}
\description{
Update batch state with new responses
}
\keyword{internal}
