% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellmer.R
\name{process_chunks}
\alias{process_chunks}
\title{Process chunks of prompts in parallel}
\usage{
process_chunks(chunks, result, chat_obj, type_spec, pb, state_path, echo, beep)
}
\description{
Process chunks of prompts in parallel
}
\keyword{internal}
