% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellmer.R
\name{process_parallel}
\alias{process_parallel}
\title{Process batch of prompts in parallel}
\usage{
process_parallel(
  chat_obj,
  prompts,
  type_spec = NULL,
  state_path = tempfile("chat_batch_", fileext = ".rds"),
  chunk_size = 4,
  workers = 4,
  echo = "none",
  beep = TRUE,
  parallel_plan = "multisession"
)
}
\description{
Processes a batch of chat prompts using parallel workers.
}
\keyword{internal}
