% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{progress.batch}
\alias{progress.batch}
\title{Extract progress information from a batch}
\arguments{
\item{x}{A batch object}
}
\value{
A list containing progress details
}
\description{
Extract progress information from a batch
}
