% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{progress}
\alias{progress}
\title{Get progress information from a batch result}
\usage{
progress(x, ...)
}
\arguments{
\item{x}{A batch object}

\item{...}{Additional arguments passed to methods}
}
\value{
A list containing progress details
}
\description{
Get progress information from a batch result
}
