% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{structured_data.batch}
\alias{structured_data.batch}
\title{Extract structured data from a batch}
\arguments{
\item{x}{A batch object}
}
\value{
List of structured data
}
\description{
Extract structured data from a batch
}
