% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{structured_data}
\alias{structured_data}
\title{Extract structured data from a batch result}
\usage{
structured_data(x, ...)
}
\arguments{
\item{x}{A batch object}

\item{...}{Additional arguments passed to methods}
}
\value{
A list of structured data objects
}
\description{
Extract structured data from a batch result
}
