% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellmer.R
\name{texts.batch}
\alias{texts.batch}
\title{Extract text responses from a batch}
\arguments{
\item{x}{A batch object}
}
\value{
Character vector of response texts
}
\description{
Extract text responses from a batch
}
