% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellmer.R
\name{update_batch_state}
\alias{update_batch_state}
\title{Update batch state with new responses}
\usage{
update_batch_state(result, new_responses, pb, state_path)
}
\description{
Update batch state with new responses
}
\keyword{internal}
