% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellmer.R
\name{batch}
\alias{batch}
\title{Batch class for managing chat processing}
\usage{
batch(
  prompts = list(),
  responses = list(),
  completed = integer(0),
  state_path = character(0),
  type_spec = NULL,
  echo = character(0),
  progress_bar = NULL,
  input_type = character(0)
)
}
\description{
Batch class for managing chat processing
}
