% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellmer.R
\name{chat_batch}
\alias{chat_batch}
\title{Process batch of prompts}
\usage{
chat_batch(chat_model = chat_openai(), echo = "text", ...)
}
\arguments{
\item{echo}{Echo level}

\item{chat_obj}{Chat model object}

\item{prompts}{List of prompts}

\item{type_spec}{Type specification for structured data}

\item{state_path}{Path for saving state}
}
\value{
Batch results object
}
\description{
Process batch of prompts
}
