% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{chats}
\alias{chats}
\title{Extract chat objects from a batch result}
\usage{
chats(x, ...)
}
\arguments{
\item{x}{A batch object}

\item{...}{Additional arguments passed to methods}
}
\value{
A list of chat objects

A list of chat objects
}
\description{
Extract chat objects from a batch result

Extract chat objects from a batch result
}
