% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellmer.R
\docType{package}
\name{hellmer-package}
\alias{hellmer}
\alias{hellmer-package}
\title{hellmer: Batch Processing for Chat Models}
\description{
This package enables batch processing of chat models from 'ellmer' (\url{https://github.com/tidyverse/ellmer}) supporting sequential or parallel processing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dylanpieper.github.io/hellmer/}
}

}
\author{
\strong{Maintainer}: Dylan Pieper \email{dylanpieper@gmail.com}

}
\keyword{internal}
