% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-classes.R
\docType{package}
\name{hellmer-package}
\alias{hellmer}
\alias{hellmer-package}
\title{hellmer: Batch Processing for Chat Models}
\description{
This package enables sequential or parallel batch processing for chat models from 'ellmer'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dylanpieper.github.io/hellmer/}
}

}
\author{
\strong{Maintainer}: Dylan Pieper \email{dylanpieper@gmail.com}

}
\keyword{internal}
