% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellmer.R
\docType{package}
\name{hellmer-package}
\alias{hellmer}
\alias{hellmer-package}
\title{hellmer: Batch Processing for Chat Models}
\description{
This package implements batch processing for chat models from the 'ellmer' package (\url{https://github.com/tidyverse/ellmer}).
}
\author{
\strong{Maintainer}: Dylan Pieper \email{dylanpieper@gmail.com}

}
\keyword{internal}
