% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{texts.batch}
\alias{texts.batch}
\title{Extract text responses or structured data from a batch}
\arguments{
\item{x}{A batch object}

\item{flatten}{Logical; whether to flatten structured data into a single string (default: TRUE)}
}
\value{
A character vector (if original prompts were supplied as a vector) or
a list of response texts (if original prompts were supplied as a list).
If a type specification was provided to the batch, structured data objects will be returned instead.
}
\description{
Extract text responses or structured data from a batch
}
