% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellmer.R
\name{texts.batch}
\alias{texts.batch}
\title{Extract text responses from a batch}
\arguments{
\item{x}{A batch object}
}
\value{
A character vector (if original prompts were supplied as a vector) or
a list of response texts (if original prompts were supplied as a list)
}
\description{
Extract text responses from a batch
}
