% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-classes.R
\name{texts}
\alias{texts}
\title{Extract texts from a batch result}
\usage{
texts(x, ...)
}
\arguments{
\item{x}{A batch object}

\item{...}{Additional arguments passed to methods}
}
\value{
A character vector or list of text responses
}
\description{
Extract texts from a batch result
}
